Allow separate service configuration for connectors.
For example, a FacebookPage connector cannot instantiate a
FacebookPage service from PHPoAuthLib because it doesn't exist,
so it should use the Facebook service instead. For that to happen,
the FacebookPage key in the configuration array (which is itself another
array) should have a 'service' key, defining which PHPoAuthLib service
to use.